requirements.txt
anthropic==0.79.0
flask==3.1.2
gunicorn==23.0.0
python-dotenv==1.2.1
